batch/requirements.txt
streamlit
pandas
openpyxl
python-docx
Pillow
reportlab
marker-pdf
torch
torchvision
torchaudio
pypdf2